requirements.txt
pyyaml>=6.0
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pgvector==0.3.6
openai==1.59.5
anthropic==0.42.0
beautifulsoup4==4.12.3
requests==2.32.3
youtube-transcript-api==0.6.3
pydantic==2.10.4
pydantic-settings==2.7.1
tiktoken==0.8.0
numpy==2.2.1
python-dotenv==1.1.0
sentence-transformers==3.4.1
scikit-learn==1.6.1
httpx==0.28.1
fastmcp==2.3.4
google-generativeai==0.8.3
